Finish getLunch function to generate lunch group for user
Modified getLunch function both in views.py and models.py to generate random colleagues for a user. Added several edge cases to the getLunch function in models that handles cases such as:1 user in the system, less than 3 users in the system. It also prioritizes friends not met for lunch before while finding a group for lunch.
Also added an addLunchFriend function that adds new friends to user's lunch_firnds list.